Coding projects/PDF article manager/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-article-manager"
version = "0.1.0"
description = "Scan, classify, and organize scientific article PDFs using AI"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Elixr Bio"}
]
dependencies = [
    "PyMuPDF>=1.24",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "tqdm>=4.66",
]

[project.scripts]
pdf-manager = "pdf_manager.__main__:main"

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["src"]
